requirements.txt
pandas
plotly
plotly.express
seaborn
matplotlib
streamlit
joblib
numpy
Pillow
scikit-learn
